+++ python/beeai_framework/agents/requirement/utils/_tool.py
@@ -4,15 +4,17 @@
 import asyncio
 import json
 from asyncio import create_task
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, Self
 
 from pydantic import BaseModel, Field, InstanceOf
 
 from beeai_framework.backend import AssistantMessage, MessageToolCallContent
+from beeai_framework.backend.errors import ChatModelToolCallError
 from beeai_framework.context import RunContext
 from beeai_framework.emitter import Emitter
 from beeai_framework.errors import FrameworkError
 from beeai_framework.tools import AnyTool, StringToolOutput, Tool, ToolError, ToolOutput, ToolRunOptions
+from beeai_framework.utils.strings import to_json
 
 if TYPE_CHECKING:
     from beeai_framework.agents.requirement import RequirementAgentRunState
@@ -23,6 +25,12 @@ async def _run_tool(
     msg: MessageToolCallContent,
     context: dict[str, Any],
 ) -> "ToolInvocationResult":
+    if not msg.is_valid():
+        raise ChatModelToolCallError(
+            generated_content=to_json({"name": msg.tool_name, "parameters": msg.args}, sort_keys=False),
+            generated_error="The generated tool call is invalid. Cannot parse the args.",
+        )
+
     result = ToolInvocationResult(
         msg=msg,
         tool=None,
@@ -96,6 +104,14 @@ class FinalAnswerTool(Tool[BaseModel, ToolRunOptions, StringToolOutput]):
             self._state.answer = AssistantMessage(input.response)  # type: ignore
 
         return StringToolOutput("Message has been sent")
+
+    async def clone(self) -> Self:
+        tool = self.__class__(expected_output=self._expected_output, state=self._state.model_copy())
+        tool.name = self.name
+        tool.description = self.description
+        tool._cache = await self.cache.clone()
+        tool.middlewares.extend(self.middlewares)
+        return tool
 
 
 class ToolInvocationResult(BaseModel):