+++ python/beeai_framework/utils/models.py
@@ -6,7 +6,6 @@ import copy
 from abc import ABC
 from collections.abc import Generator, Sequence
 from contextlib import suppress
-from logging import Logger
 from typing import Any, Generic, Literal, Optional, Self, TypeGuard, TypeVar, Union
 
 from pydantic import BaseModel, ConfigDict, Field, GetJsonSchemaHandler, RootModel, ValidationError, create_model
@@ -16,8 +15,6 @@ from pydantic_core import CoreSchema, SchemaValidator
 
 from beeai_framework.utils.dicts import remap_key
 from beeai_framework.utils.schema import simplify_json_schema
-
-logger = Logger(__name__)
 
 T = TypeVar("T", bound=BaseModel)
 ModelLike = Union[T, dict[str, Any]]  # noqa: UP007
@@ -105,10 +102,6 @@ class JSONSchemaModel(ABC, BaseModel):
             )
 
             if one_of:
-                logger.debug(
-                    f"{JSONSchemaModel.__name__}: does not support 'oneOf' modifier found in {param_name} attribute."
-                    f" Will use 'anyOf' instead."
-                )
                 return create_field(param_name, remap_key(param, source="oneOf", target="anyOf"), required)
 
             if any_of:
@@ -152,10 +145,6 @@ class JSONSchemaModel(ABC, BaseModel):
                 if isinstance(const, str):
                     target_type = Literal[const]
                 if not target_type:
-                    logger.debug(
-                        f"{JSONSchemaModel.__name__}: Can't resolve a correct type for '{param_name}' attribute."
-                        f" Using 'Any' as a fallback."
-                    )
                     target_type = type
 
                 if (not is_required and not default) or explicitly_nullable: